+++ packages/spacecat-shared-utils/src/index.js
@@ -31,6 +31,8 @@ export {
   isValidIMSOrgId,
   isValidHelixPreviewUrl,
   toBoolean,
+  UUID_REGEX,
+  UUID_V4_REGEX,
 } from './functions.js';
 
 export { isValidEmail } from './email.js';