+++ packages/spacecat-shared-splunk-client/CHANGELOG.md
@@ -1,3 +1,22 @@
+## [@adobe/spacecat-shared-splunk-client-v1.2.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-splunk-client-v1.2.0...@adobe/spacecat-shared-splunk-client-v1.2.1) (2026-05-29)
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#1609](https://github.com/adobe/spacecat-shared/issues/1609)) ([ccdcc2b](https://github.com/adobe/spacecat-shared/commit/ccdcc2b88c1b03e5c58ed682daaaadb382638395))
+* **deps:** update external major (major) ([#1611](https://github.com/adobe/spacecat-shared/issues/1611)) ([5a80797](https://github.com/adobe/spacecat-shared/commit/5a807979187f1ebcc32b088faee9165aa38cc461))
+
+## [@adobe/spacecat-shared-splunk-client-v1.2.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-splunk-client-v1.1.6...@adobe/spacecat-shared-splunk-client-v1.2.0) (2026-05-21)
+
+### Features
+
+* **ci:** migrate npm publishing to OIDC Trusted Publishers ([#1592](https://github.com/adobe/spacecat-shared/issues/1592)) ([7556295](https://github.com/adobe/spacecat-shared/commit/75562957a223b5783a97998c8dc12eebd3e4db34))
+
+## [@adobe/spacecat-shared-splunk-client-v1.1.6](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-splunk-client-v1.1.5...@adobe/spacecat-shared-splunk-client-v1.1.6) (2026-05-14)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1533](https://github.com/adobe/spacecat-shared/issues/1533)) ([0a3e2ab](https://github.com/adobe/spacecat-shared/commit/0a3e2abbbc5f58b5320518f7d596d4cef6271fa0))
+
 ## [@adobe/spacecat-shared-splunk-client-v1.1.5](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-splunk-client-v1.1.4...@adobe/spacecat-shared-splunk-client-v1.1.5) (2026-04-09)
 
 ### Bug Fixes